Create Rock, Paper, Scissors Game
Created a function to generate a random computer choice.
Created a function to prompt the user for an input.
Created a function to play a round of the game and decide who wins.
Created a function to iterate 5 rounds and declare an overall winner.

Currently I can get the game to work but only with an extra prompt?
Needs more work.